+++ packages/sdk/openfeature-node-server/src/LaunchDarklyProvider.ts
@@ -24,7 +24,7 @@ export default class LaunchDarklyProvider extends BaseOpenFeatureProvider<LDClie
       const client = init(sdkKey, {
         ...options,
         wrapperName: 'open-feature-node-server',
-        wrapperVersion: '1.3.6', // x-release-please-version
+        wrapperVersion: '2.0.0', // x-release-please-version
       });
 
       this.setClient(client);