+++ packages/sdk/vue/src/index.ts
@@ -22,6 +22,7 @@ export type {
   LDFlagValue,
   LDInspection,
   LDLogger,
+  LDPlugin,
   Hook,
   LDIdentifyOptions,
   LDIdentifyResult,