+++ packages/sdk/akamai-edgekv/CHANGELOG.md
@@ -31,6 +31,25 @@ All notable changes to the LaunchDarkly SDK for Akamai Workers will be documente
     * @launchdarkly/akamai-edgeworker-sdk-common bumped from ^1.1.1 to ^1.1.2
     * @launchdarkly/js-server-sdk-common bumped from ^2.2.1 to ^2.2.2
 
+## [1.4.33](https://github.com/launchdarkly/js-core/compare/akamai-server-edgekv-sdk-v1.4.32...akamai-server-edgekv-sdk-v1.4.33) (2026-08-25)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/akamai-edgeworker-sdk-common bumped from ^2.0.29 to ^2.0.30
+
+## [1.4.32](https://github.com/launchdarkly/js-core/compare/akamai-server-edgekv-sdk-v1.4.31...akamai-server-edgekv-sdk-v1.4.32) (2026-08-18)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/akamai-edgeworker-sdk-common bumped from ^2.0.28 to ^2.0.29
+    * @launchdarkly/js-server-sdk-common bumped from ^2.20.0 to ^2.20.1
+
 ## [1.4.31](https://github.com/launchdarkly/js-core/compare/akamai-server-edgekv-sdk-v1.4.30...akamai-server-edgekv-sdk-v1.4.31) (2026-07-28)