+++ packages/sdk/react/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## [4.1.14](https://github.com/launchdarkly/js-core/compare/react-sdk-v4.1.13...react-sdk-v4.1.14) (2026-08-26)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/js-client-sdk bumped from ^4.9.4 to ^4.10.0
+
+## [4.1.13](https://github.com/launchdarkly/js-core/compare/react-sdk-v4.1.12...react-sdk-v4.1.13) (2026-08-18)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/js-server-sdk-common bumped from ^2.20.0 to ^2.20.1
+
 ## [4.1.12](https://github.com/launchdarkly/js-core/compare/react-sdk-v4.1.11...react-sdk-v4.1.12) (2026-08-07)