+++ packages/sdk/server-node/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to `@launchdarkly/node-server-sdk` will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org).
 
+## [9.13.1](https://github.com/launchdarkly/js-core/compare/node-server-sdk-v9.13.0...node-server-sdk-v9.13.1) (2026-08-18)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/js-server-sdk-common bumped from 2.20.0 to 2.20.1
+
 ## [9.13.0](https://github.com/launchdarkly/js-core/compare/node-server-sdk-v9.12.3...node-server-sdk-v9.13.0) (2026-07-28)