+++ packages/sdk/vue/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [0.2.2](https://github.com/launchdarkly/js-core/compare/vue-client-sdk-v0.2.1...vue-client-sdk-v0.2.2) (2026-08-28)
+
+
+### Bug Fixes
+
+* **vue-sdk:** re-export `LDPlugin` from base impl ([#1897](https://github.com/launchdarkly/js-core/issues/1897)) ([e21f484](https://github.com/launchdarkly/js-core/commit/e21f4842f2f1f5e3223e35e26cb82bb171681f2f))
+
+## [0.2.1](https://github.com/launchdarkly/js-core/compare/vue-client-sdk-v0.2.0...vue-client-sdk-v0.2.1) (2026-08-26)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/js-client-sdk bumped from 4.9.4 to 4.10.0
+
 ## [0.2.0](https://github.com/launchdarkly/js-core/compare/vue-client-sdk-v0.1.4...vue-client-sdk-v0.2.0) (2026-08-14)