+++ packages/tooling/jest/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [1.0.27](https://github.com/launchdarkly/js-core/compare/jest-v1.0.26...jest-v1.0.27) (2026-08-26)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/react-native-client-sdk bumped from ~10.19.4 to ~10.20.0
+
 ## [1.0.26](https://github.com/launchdarkly/js-core/compare/jest-v1.0.25...jest-v1.0.26) (2026-08-06)